+++ kitty/boss.py
@@ -3402,4 +3402,6 @@ class Boss:
     def ungrab_keyboard(self) -> None:
         grab_keyboard(False)
 
-
+    def search_scrollback_in_active(self) -> None:
+        if w := self.active_window:
+            w.search_scrollback()